My Notes by Xamarin.Forms
It's very common for a mobile application to use a local SQLite
database. The combination of Sqlite.net and Xamarin.Forms makes this
very easy. One of the reasons that we choose SQLite as our mobile
database is that it's a single file and easily works across platforms.
iOS and Android both have the SQLite database engine "built in” and
access to the data is simplified by Xamarin's platform that comes with
the SQLite Data Provider.

Here I'm showing how to create a small project to store my notes into
the local database. In this app we can save data, view the saved data
and also we can update and delete the saved data.